Split up CellParser into Convertible
